CaWE: Make sure that entities are saved in the `cmap` file in the same (depth-first) order as in the `cent` file.
The map loading code relies on the depth-first ordering of the entities both in
the `cmap` and the `cent` file, but with the previous code this only held when
the root entity was the *only* entity with children, i.e. the depth of the
entity hierarchy was not greater than 1.