CaWE: Convert `func_door` entities to "door with parts" entity hierarchies.
Each "team" of `func_door` entities is replaced with a single new entity that
has a script component, and each former team member is now kept as a child
entity of the new "main" door entity.

As a result, door functionality is now broken, because this is the first time
that an entity other than the root entity has children. This in turn requires
proper dealing with model-to-world and world-to-model transforms, which many
parts of our programs are not yet (fully or properly) aware of.

This change was intentionally made to trigger all these problems, and it is
subject to the next revisions to fix all their occurrences.