CaBSP: Mildly improve the transform of map primitives from world to model space.
In CaWE, all map primitives are kept and saved to disk in world space at this
time. Therefore, when CaBSP initially loads them, it transforms the map
primitives from world space to the local space of their entity.
This transform was never aware of entity hierarchies and of rotations, though.

This change makes it aware of entity hierarchies, but not of rotations.
Please see the comment in the change itself for details.