GameSys: Clarify world-space vs. parent-space vs. entity-space.
Although this distiction was always clear in the Transform component (see the
documentation of the `ComponentTransformT` class for details), user code
historically often assumed that "everything" was in world-space anyway, and
thus was agnostic of the space the coordinates were really in.

As we now start using non-trivial entity hierarchies for the first time, this
change improves the Transform component's API to make it explicitly clear to
user code which coordinate space it is dealing with.