DeathMatch: Remove the old game code's knowledge of a map's precomputed collision models.
Collision models for map entities that are composed of map primitives have been
handled by the Cafu Engine's "Application" component for quite a while, and the
old DeathMatch game code just uselessly duplicated the functionality.
As the latter was no longer properly updated when the related entity moved
(e.g. a rigid body or an opening door) anyway, effectively blocking the player's
movement where no blocking was due, removing the offending and unneeded code was
the obvious step.