DeathMatch: Have the old game code "sync" the human player origin with its new Transform component.
At this time, human player computations are still in "old-style" code, but for
the upcoming CompanyBot re-implementation, we need to know the human player's
origin in "new-style" code (the entity's Transform component) as well.

This is just some glue code to bridge the time until human player entities are
re-imlemented in the terms of the component system as well.