GameSys: Replace the Model component's `SetAnim()` script method with the new `OnAnimationChange()` callback.
When `SetAnim()` was called, it was usually from the `Think()` method of a
Script component, and thus on the *server* side only. Thus, the parameters
passed to `SetAnim()` were only known to and used on the server, but not on the
client, where they are really wanted and needed.

The new `OnAnimationChange()` callback is more generic, as the code calls it
whenever a new animation sequence is set, allowing its implementation to
"configure" the new sequence, no matter if on the client or the server.
(This should later be expanded to deal with animation "channels" as well.)