GameSys: Have PlayerPhysics components ignore the first CollisionModel in the same entity.
If the PlayerPhysics component doesn't explicitly ignore the clip models of a
CollisionModel, its traces all start in solid and thus become meaningless.

At this time, only the first CollisionModel component in the entity is
accounted for. This should possibly be expanded to arbitrarily many
CollisionModels and possibly even to the Cafu Engine's "Application" component.