Engine: Properly account for new entities that have been created in a `Think()` step by script code (`ComponentScriptT`).
If new-style component-system script code created new entities in a `Think()`
step, the new entities exist in the script world, but not yet in the server's
`m_EngineEntities` array, and they don't have an "App" component yet. We thus
monitor the list of entities for such cases and take the appropriate steps.