GameSys: Add new component `ComponentPlayerStartT`, and re-implement "info_player_start" entities in its terms.
It would have been possible to implement the same functionality with the
existing Script component alone, that is, have each "info_player_start" entity
have a Script component that implements `IsSinglePlayerStart()` and
`IsMultiPlayerStart()` callbacks, returning `true`.

However, as we may wish to
  - convert such entities into "prefabs" later,
  - display them prominently in the map editor CaWE later
    (both in the 2D and 3D views and in the Entity Inspector window),
  - identify them reliably in the map compile tools and human player code,
it seemed wiser to invent an explicit new component for them now.