General: Let each human player entity have a child entity that defines the first person view's "camera".
The "camera" entity's Transform component is used to setup the first person
view's origin and orientation, which can now be positioned independently of the
related human player e.g. for slight head movements ("swaying"), for smoothing
the movement when climbing steps, etc.

This part of the implementation focuses mostly on the Engine side's required changes.
Updating the human player entity code accordingly is subject of the next revision.