General: Have camera entities look along their local x-axis rather than the y-axis.
In the OpenGL default coordinate system, the camera looks along the negative
z-axis.

In Cafu, we used to have it look along the (positive) y-axis, which was the
loose but long-time convention (I even explicitly referred to it only three
revisions ago in 2efa206).

However, it turns out that we have to have it look along the (positive) x-axis
in order to have it comply to the conventions defined with the
`cf::math::AnglesT<T>` class. Only then can we meaningfully use the angles in
this class in order to get and set the orientation of camera entities.