DeathMatch: Transfer all physics (rigid body) handling of human player entities to the component system.
This change actually *removes* the rigid body from the `EntHumanPlayerT` class.
It also updates all code (especially the carried weapons' tracing of rays) to
account for the Physics components of human player entities in the new
component system.

However, it does *not* actually add Physics components to human player entities,
as would be required to compensate for the removal in the old `EntHumanPlayerT`!

This is a task that must still be done later, because

  a) the Physics component handling depends on the "state of existance" of the
     player ("active" if alive, "not active" if dead, spectating, etc.), and the
     `Think()` code, at this time still in `EntHumanPlayerT`, must/should be
     ported to the Component System first;

  b) the Physics component code must be updated to be able to deal with
     "kinematic" bodies, and to have custom collision shapes set.