GameSys: Make it possible to post and process events through the HumanPlayer component.
This is achieved by "routing" the posting and processing of events via the
Script component, which actually provides all functionality.

The only purpose of these measures is to further facilitate the `CarriedWeaponT`
method implementations, especially during the transition from the old DeathMatch
game code to the entity component system. We may undo the "routing" introduced
in this commit if we can come up with a more refined solution later.