DeathMatch: Replace the rests of the `ResourceManagerT` with an interim solution.
The interim solution is transparent to users, but should be replaced, along with
`ComponentParticleSystemOldT`, with something better as soon as the transition
to the component system is complete.

The actual purpose of this change is to get rid of the dependencies of the
`CarriedWeaponT` classes on the `ResourceManagerT` class and thus on the old
`GameImplT`.