GameSys, DeathMatch: Move the list of `CarriedWeaponT` instances from the old `GameImplT` into the new `cf::GameSys::WorldT`.
This is another important step towards implementing the human player entity in
terms of the new component system.

Note however that the "carried weapons" code does currently not match the
spirit and philosophy of the component system, and that we thus will later have
to revise the code so that it is possible to define "carried weapons" just like
any other component.