GameSys: In the HumanPlayer component, re-enable the processing of incoming events by the "carried weapons" code.
Incoming events were properly received by the HumanPlayer component since commit
ed678dc, but could not be forwarded from there to the active carried weapon as
long as the `CarriedWeaponT` instances were still in the old DeathMatch code.
As this has been rectified in the previous commit, it was now possible to re-
establish the proper functionality.