GameSys: Teach the `CarriedWeaponT` classes to deal with `NULL` instances of the global `MatSys::Renderer` and `SoundSystem`.
As the `CarriedWeaponT` instances now live in the new `cf::GameSys::WorldT`
(see commit 5097c3e), they may also be instantiated in the Map Editor, the map
compile tools, and the server, where at this time, the `SoundSystem` and/or
the `MatSys::Renderer` can be `NULL`.

This change updates the `CarriedWeaponT` classes and the ParticleEngine to
deal with the situation. Independently of that, both the `CarriedWeaponT`
classes and the ParticleEngine remain in urgent need of a thorough overhaul.