DeathMatch: Delete the no longer needed `EntHumanPlayerT` class, all whose functionality has now been re-implemented in terms of the entity component system!
At this time, the re-implementation in the HumanPlayer component is not perfect
and not as polished as the `EntHumanPlayerT` used to be, as a few compromises
had to be made. However, it is expected that with the progressing and almost
complete removal of the obsolete DeathMatch code it becomes possible to polish
the new code to at least the same shininess as the previous code, which will
thoroughly be done.

This change is a very important milestone and huge symbolic step in the
introduction of the entity component system!   :-)