Engine: Remove a large amount of old and no longer needed cruft.
WOW! Finally!!  :-)

This is a very important milestone in the introduction of the entity component
system, as this change makes it unequivocally clear that it is the new entity
component system code, and the new entity component system code only, that
implements and runs our games.

The old code was also complex and intertwined. Removing it makes reading and
maintaining the remaining good code a lot clearer and more enjoyable.