Engine: Done -- remove the last remaining fragments of the old DeathMatch C++ game code!
The important feature of this change is that it is now no longer necessary to
write any custom C++ code to create a new game with the Cafu Engine.

The new entity component system code has overtaken all functionality that the
old game code still had, and all features of the Cafu Engine are provided to
game developers via the components and their support for scripting.

Note however that the introduction of the entity component system is, at this
point, by no means complete: A lot of code is still to be fixed, improved and
polished, and in the CaWE Map Editor even more work is to be done.

Nevertheless, we will now no longer be removing old and obsolete code, but
build and improve upon the new and shiny component system code that has been
carefully crafted and added over the past couple of months.