GameSys: For each human player entity, add another child entity that holds the 1st-person weapon models.
This change builds on recent revision c97a2f7. At this time, its functionality
is quite isolated, because it only "duplicates" the 1st-person weapon model,
but lacks all related wiring. (For example, its `OnSequenceWrap()` script
callback should be implemented, as well as any other related functionality.)

The key idea is to use this as a first step to re-implement all carried weapons
code anew, using components, in parallel to the old, existing code. When the
new code is on par with the old code, the old code is removed and the new code
takes its place.