Network: Increase the maximum packet size that we can send over the network.
The previous value was set low in order to avoid fragmentation at the IP layer,
but becomes increasingly problematic with entities that have many components.
(Actually the upcoming changes to human player entities are the reason for this
change.)

Note that in mid-term, it is intended to replace our own UDP-based game protocol
implementation with SCTP, which will also help with dealing with packets that
occasionally exceed the normally expected size.