GameSys: Introduce a set of C++ methods and Lua script callbacks that work together so that (human player) entities can select (holster and draw) among their available weapons.
This is a big change that establishes the core concepts of carried weapons
handling. The key idea is that the CarriedWeapon components themselves are
responsible for managing the weapon state and the related 1st-person model.

As a result, the player can now pick up weapons from the ground and manually
toggle among them, where the holstering and drawing is properly managed.

For a detailed description, see the newly added "Carried Weapons Overview" in
the scripting reference documentation (available online at
http://docs.cafu.de/lua/CarriedWeaponsOverview.html).