CaWE: Undertake a search for the Cafu executables if none are configured.
When CaWE is newly installed or for any other reasons no paths to the Cafu
executables are configured, CaWE used to use plain default values that were
tailored to our binary releases. However, this simple strategy failed whenever
Cafu was compiled from source code, requiring the user to manually setup the
proper paths.

With this change, we try to determine the proper paths automatically, using the
same "search algorithm" that is used in the (otherwise unrelated)
`Games/DeathMatch/compileMaps.py` script. Only if this step fails (e.g. because
it is *really* a binary release) are the old default values used again.

Closes #132.