GameSys: In the HumanPlayer component, have weapon index numbers start at 1 (rather than 0).
This is done so that 0 can be used to unambiguously refer to "none", i.e.,
"no active weapon". Without this, an "invalid" value such as 255 had to be used
for the same purpose, which is not unique (254 is "invalid" as well) and
difficult to communicate across both C++ and Lua.