GameSys, DeathMatch: Move the names of the 1st- and 3rd-person weapon models into the weapon scripts.
It makes no sense to keep these names as string members in the CarriedWeapon
components, because a weapon model can normally not be changed without careful
updates to and revision of the related weapon script anyway.
Thus, keeping the names as members in the component just increases complexity
and wastes bandwidth.

(Note that it would also be possible to remove the members for the "maximum"
amounts of primary and secondary ammo in a similar manner...)