GameSys, DeathMatch: Add new method `RegisterParticle()` to the HumanPlayer component, and use it to implement the remaining particle effects of the weapons.
The new method is another clumsy hack, and intended to be replaced with a better
approach later (the entire Particle Engine needs a thorough overhaul).
However, it allows us to complete the implementation of the weapons in Lua,
and to remove their old implementation in C++ in one of the next commits.