Map Editor: Move the `CommandHistoryT` instance from the `MapDocumentT` into the `ChildFrameT`.
The `ChildFrameT` instance is where the `CommandHistoryT` should have always
been, as it logically is something "external" that modifies the `MapDocumentT`,
and thus cannot be a part of it.
This change makes the setup analogous to the GUI and Model Editors.

The change also required the move of some menu item event handler methods,
but the new place is where they too should have always been anyway.