Map Editor: Revise the `Insert()` and `Remove()` methods of the `MapDocumentT`.
The methods now work recursively, i.e. they properly account for all of the
entity's children when updating the `MapDocumentT`'s underlying BSP tree.

The `Insert()` method now also accepts `Parent` and `Pos` parameters, so that
a better control of where exactly a new entity is inserted is possible for the
caller. (The "Undo" action of the "Delete" command can use that to properly
restore an entity's position in the entity hierarchy, see next commit.)