Map Editor: In the Entity Hierarchy, use labels of the form "Entity name (number of primitives)".
The goal of this change is to provide the user with a better sense of how the
primitives relate to their respective parent entities.

The details of the implementation are a bit awkward, because we have to "tweak"
the label before and after the label is edited (renamed). Using a
`wxTreeListCtrl` would have been the cleaner solution, but it would also have
required a lot more work and the new column for the number of primitives would
have consumed much more screen space.