Map Editor: Revive the "auto-group entities" feature, with which all primitives and children of an entity can be selected as a group.
This is only a quick temporary measure. The better solution is to integrate the
"Groups" feature into the entities themselves (that is, into the
Entity Hierarchy dialog and the `CompMapEntityT` component), which is possible
now that entities are of a much different nature than they were in
pre-component-system times. Implementing this solution is a bigger change
though, and it must be implemented separately in future commits.