Map Editor: Remove the "Assign primitive to world" special-case menu item and toolbar button.
In pre-component-system times, the world (root) entity was implicit and
"invisible" to the user -- it could not explicitly be selected and worked with.
Now, with the component-system, the world (root) entity is explicit and can be
selected and used like any other entity.

Therefore, special-case treatment of the world entity, as the
"Assign primitive to world" special-case menu item and toolbar button were
instances of, is no longer needed: all functionality is covered by the general-
purpose "Assign primitive to entity" code.