Map Editor: Revise the handling, rendering and use of bounding-boxes of `MapEntRepresT`s.
This is a big improvement, because the `MapEntRepresT` implementation now
internally discriminates between "culling" and "editor handle" bounding-boxes.

It is still not perfect though, because the public interface only exposes the
`GetBB()` method, so that external user code is still forced and limited to
using only a single bounding-box for all purposes.