Map Editor: If the entity hierarchy changes, keep the rearranged entity's world-space transform unchanged.
If we drag-and-dropped an entity into a new parent, the entity's transform,
which is in parent-space, caused the entity to get a new position and
orientation in world-space, because with the parent also the parent's new
transform was adopted.

This change makes sure that the entity's (parent-space) transform is updated so
that its world-space transform remains unchanged when the entity is assigned
a new parent.