Map Editor: Add a new wxDataViewCtrl-based implementation for the "Entity Hierarchy" dialog.
The wxDataViewCtrl is unfortunately not unproblematic, it requires the changes
to wxWidgets submitted in cfd946f to work.

The new icons in the rightmost columns are at this time not yet implemented,
they're only for mockup.

Also, at this time both the "old" and "new" implementations of the
"Entity Hierarchy" dialog co-exists: both are visible, both work, and both are
independent of each other. The "old" wxTreeCtrl-based implementation will
however be disabled soon.