Map Editor: Render a visual representation for entities even if they have no custom components at all.
The previous code was too confusing, because newly created entities seemed to
disappear when deselected, entities that are used for grouping seemed not to
exist, etc. For such entities, we now render a simple wire-frame box in the 3D
views, which is enough to indicate their presence, but still unobtrusive enough
to not distract the user.