GameSys: Augment the Basics component with new members `Show` and `SelMode` that are specific to the Map Editor.
The Basics component had a `Show` member before, but it was a leftover from the
beginnings of the game component system when it was initially created as a copy
of the GUI component system. In a sense, in this commit the "old" `Show` member
was removed and then re-introduced with a different meaning.

Both new members are specific to the Map Editor, carrying data that should be
loaded and saved along with a map.

Keeping these new members inside the (otherwise "purely general purpose")
Basics component seems awkward at first, but it is by far the most natural
solution, considering the alternatives (e.g. keeping the same data in the
entity's "properties", which are obsolete, or introducing a new component,
which had to live along with the others in `Libs/GameSys`, but still would have
awkwardly needed some special-case code at least in the Map Editor).