Map Editor: Remove main menu items and toolbar buttons that were related to the obsolete "groups" feature.
As the "groups" feature is obsolete, several main menu items and buttons of the
Selection tool's toolbar are no longer needed:

"Group"
  : groups the currently selected objects

"Hide"
  : groups the currently selected objects, and hides the newly created group

"Hide Other"
  : groups all unselected map elements that are not in a group already,
    and hides the newly created group

"Show all"
  : un-hide all groups

These items are mirrored in the new, existing functionality (where entities are
used as natural "replacement" for groups). That is, the
"assign objects to entity" functionality and the ability to show/hide entire
entities can be used to achieve the same effects as the above remove items.

(Later commits may add convenient shortcuts though, e.g. in the right-click
context menu of the Entity Hierarchy dialog.)