Map Editor: Have `cent` map files use local variables only.
With this change, loaded map files no longer modify the global script state, so
that multiple maps can be loaded into the same world without interfering with
each other (that is, without overwriting each other's global variables).
This is especially useful with "prefabs", which we implement as small but normal
map files.

See also:

  - commit [926ac9d](https://bitbucket.org/cafu/cafu/commits/926ac9d2dce01e7c3f1c2457e34ae46553bd711b)
    of 2013-11-28, where the same was introduced for GUI scripts,
  - forum post http://www.cafu.de/forum/viewtopic.php?p=6456#p6456