GameSys: Augment the `WorldT::LoadScript()` function so that it can be used several times, especially for post-loading other maps (prefabs) into existing maps.
Post-loading is somewhat "specialized" to loading prefabs, because regular maps
usually have `cw` world files associated to them, which in turn imply certain
requirements on the IDs of the entities. These requirements cannot be met when
maps are "post-loaded", and thus post-loading maps aka. "prefabs" requires at
least a small amount of special-case treatment.

Moreover, all map scripts are expected to call the `world:SetRootEntity()`
script method, but post-loaded prefabs must not change the root entity of the
map they are loaded into. This requires some special-case treatment as well.