Map Editor: Properly transform the primitives of newly loaded prefabs.
In a prefab's `cmap` file, the primitives are defined in "prefab world-space",
which corresponds to the local space of the entity that the prefab will be added
to as a child, that is, it corresponds to "parent-space".

This change transforms the primitives of a prefab from parent-space to their
proper position in the map they are inserted into in world-space, as defined
by the local space of the parent entity they're attached to.

As a second feature, this change also accounts for a translation offset that is
applied to the newly inserted prefabs as described in the source code comment.