Map Editor: Have the Entity Inspector better deal with selections that contain multiple entities.
If more than one entity is selected, and one of the selected entities is a
parent of all other selected entities, the Entity Inspector will now show the
details of the selected common parent (rather than the barely helpful
"multiple selection" notice).

This improves the usability of the Map Editor whenever an entity and all of its
children are simultaneously selected.