Map Editor: Revise some of the Tools menu's hotkeys.
Previously, some of the hotkeys also mapped to printable characters, e.g.
`Shift+S`. This in turn made it impossible to enter a capital "S" into a text
input field anywhere in the application.

See http://www.cafu.de/forum/viewtopic.php?f=9&t=1655 for a related discussion.

The root of the problem is at least partially related to the way how menu
accelerators are implemented in wxWidgets. This has been discussed several
times on the wxWidgets mailing-lists, see for example the thread at
http://thread.gmane.org/gmane.comp.lib.wxwidgets.devel/131663 for an overview.