CaWE: Add the test about proper TypeSys variable names also to the SaveComponents() function of maps and GUIs.
This is the same test about proper TypeSys variable names as previously added
to WriteDoxyVars() in `AppCaWE.cpp`. It is also added to the SaveComponents()
function of maps and GUIs because there it is run much more frequently and thus
will catch any problems much sooner than the code in WriteDoxyVars().