TypeSys: Add new method `VarManT::AddAlias()`.
The new method adds an alias name for the given variable so that a call to
`VarManT::Find()` will also find the variable under the alias name.
The purpose of this method is to provide backwards-compatibility if variables
must be renamed after they have been introduced and became widely used,
e.g. in custom user scripts.