Map Editor: Fix crash on closing a map.
We were again seeing crashes like those fixed in 294f2b7 on 2014-07-04, wherein
the cause of the problem was well documented in the comment that was added to
the `MapDocumentT`'s destructor.

It turned out that the re-occurrence of the bug was eventually related to
wxWidgets issue http://trac.wxwidgets.org/ticket/16249 (we switched to the
`wxDataViewCtrl` for displaying our Entity Hierarchy in the meanwhile), so that
the solution comprises both the application of the fix to wxWidgets and the
related update of our code.