Map Editor: Remove obsolete and problematic method `MapDocumentT::GetEntities()`.
The method was obsolete because the `GetAll()` method of the script world's
root entity is a better and more natural way to obtain the list of all entities
in a map.

The method was problematic because it contained a static array of
`IntrusivePtrT<CompMapEntityT>`s, essentially keeping `CompMapEntityT`
instances (and thus the therein contained map primitives such as Bezier
Patches) until after the end of `main()`. When these instances were eventually
released, global systems like the `MatSys::Renderer` were already unavailable,
so that releasing any related resources was no longer (properly) possible.