Map Editor: Implement a proper memento approach for encapsulating the transform-related state of a `MapElementT`.
At this time, the `CommandTransformT` and other, similar commands still rely
on the `MapElementT::Assign()` method in order to implement their `Undo()`
functionality.

This is problematic, because it handles resources wastefully, often keeping a
lot more state (clones of the `MapElementT`s) than is necessary in order to
undo a transform operation, and thus is inefficient.

This in turn is especially problematic with transforming entities:
Per definition, the Map Editor's `MapElementT`s are independent of each other,
and thus, `MapEntRepresT`s are independent of each other as well.
As a consequence, when the selected map elements are transformed, *only* the
*selected* map elements must be affected by the transform.
That especially means that if a child of an entity is *not* selected, the
child's world-space position must remain unchanged despite the transform of
its parent!
Therefore, when an entity is transformed, the transform of each child must be
updated as well, because the children's transforms are kept in parent-space,
but must remain unchanged in *world-space*.
Note that it is not necessary to consider any grand-children, as dealing with
the immediate children is sufficient.