Map Editor: Change all transform-related commands to use the new memento approach.
The previous commit introduced a proper memento approach for encapsulating the
transform-related state of a `MapElementT` for the reasons described in the
related (previous) commit message, but left the new code unused.

With this change, the transform-related commands no longer use the old, obsolete
`Clone()` and `Assign()` methods in order to implement their "undo" operation,
but instead employ the newly introduced memento approach.