Map Editor: Rewrite the material-related commands to no longer use the obsolete `MapElementT::Assign()` method for their "undo" operation.
The two previous commits "almost" allowed us to get rid of the principally
problematic `MapElementT::Assign()` method that also turned out to be very
inefficient for implementing the "undo" operation in our commands (because
the related "old state" clones of the objects contained a huge amount of
unneeded state).

This commit changes the material-related commands to no longer use the obsolete
`MapElementT::Assign()` method for their "undo" operation. As the new
implementation only saves the state that is actually needed in order to perform
the "undo" operation, as a side effect the efficiency is much improved.