Map Editor: Remove obsolete method `Assign()` from the `MapElementT` hierarchy.
As described in the previous commit messages, the `Assign()` method was both
inefficient and problematic (it is unclear what assignment means in a class
hierarchy, leaving us with unclear implementations).

The previous commits have therefore replaced all uses of the `Assign()` method
with better alternatives, so that the method is now unused and can be removed.