Map Editor: Properly transform `MapEntRepresT`s, don't inadvertently affect their children.
As already explained in greater detail in the previous commit messages,
transforming `MapEntRepresT` instances must not affect the world-space
transforms of their immediate children, because in the Map Editor, the
`MapElementT`s are independent of each other.

Before this change, transforming an entity also caused its children to be
transformed (because the transforms of the children are specified in the space
of their parent).

This was:

  - wrong for children that were *not* in the selection, because it affected
    unselected map elements, and also

  - wrong for children that *were* in the selection, because these children
    would once more be transformed "themselves", essentially undergoing the
    transform twice.