Map Editor: Improve the Selection tool's "drag-selection" preview in the 2D views.
Thanks to the recently introduced memento approach for translation-related
state, it is now possible to provide an implementation for the Selection tool's
"drag-selection" preview in the 2D views that is much simpler, more robust,
and requires fewer resources (and thus is faster).

And thanks to the fix in the previous commit, the preview is now also correct,
exactly reflecting the elements as they will be when the mouse button is
released.