Map Editor: Remove the awkward and now unused copying/cloning of `MapEntRepresT`s.
The meaning of copying or cloning a `MapEntRepresT` was always questionable,
because it is actually the `EntityT` that must be copied. Reflecting this,
the copying/cloning of `MapEntRepresT`s was awkwardly implemented.

As all code now properly copies `EntityT`s or `MapPrimitiveT`s, and no legacy
code is left that tries the copying of `MapElementT`s or `MapEntRepresT`s, the
related awkward code has been removed in this commit.