Map Editor: Improve and fix the "Copy and Paste" feature.
The "Copy and Paste" feature works "cross-document" now and uses the system
clipboard, so that the clipboard contents can also be inspected and even
modified in a text editor.

The two big problem fields for the implementation, namely properly dealing with
arbitrary hierarchies of copied objects and properly dealing with their
associated resources, have been solved well. See the implementation and the
large comment at `ChildFrameT::OnMenuEditCopy()` for details.

Limitations / TODO:

  - The initial transform of pasted objects must still be revised.
  - "Paste Special" is not yet covered.