Map Editor: Manage the "paste parent" better, making sure that "Ctrl+C, Ctrl+V" creates siblings, not descendants.
Normally, new objects are pasted "into" an entity as detailed in the comment in
`MapDocumentT::SetSelection()`. However, if an entity has just been copied into
the clipboard and is pasted briefly thereafter, the user usually doesn't expect
to find the pasted entity as a child of the source entity, but as a sibling.
Therefore, this change updates the "paste parent" record accordingly.