Map Editor: Just add a comment about `MapEntRepresT::Transform()`.
At this time, `MapEntRepresT::Transform()` is never called, because entities,
unlike primitives, are defined in parent-space whose details they carry in
their Transform component: By simply attaching an entity to a parent, it is
already "implicitly" transformed.

If `MapEntRepresT::Transform()` becomes directly used in the future, its
implementation should be updated as described in the comment.