CaWE: Use a `VarVisitorToLuaCodeT` for saving the Transform components of Map and GUI documents.
Special-case code used to be used for saving the Transform components of Map
and GUI documents. This commit changes this so that for Transform components
the same code is used as for all other components.
This was done in order to have also the numbers in the Transform components
covered by improvements to serialization as implemented in the next commit.