TypeSys: Improve `VarVisitorToLuaCodeT` to serialize floating-point variables with proper precision.
In some cases, floating-point variables used to be written with larger
precision than was necessary to reconstruct the exact floating-point value at
load time (and contrary, if we had used variables based on type `double`, the
used precision would probably have been insufficient in most cases).

Therefore, we now use the long ago introduced `serialize()` function also with
the `VarVisitorToLuaCodeT` implementation, which always picks the smallest
precision with which the exact original value can be reconstructed.