Map Editor: Revise the loading/importing of heightmaps into terrains.
The new code is more robust and provides an error message explaining the
problem rather than leaving the exception unhandled.

In the future, we should consider refactoring the whole heightmap-related code,
e.g. moving it into an explicit, commonly used `HeightmapT` class that loads,
"validates" and keeps the heightmap data from which new terrains can be
constructed or existing terrains be updated (this would also eliminate the
duplication of the `TerrainT`'s loading and validating code in `MapTerrainT`).

Fixes #145.