Map Editor: Make sure that loading prefabs or pasting the clipboard contents does not introduce new groups.
Prefabs and the clipboard can, like any regular cmap file, contain group
definitions and have their primitives associated to a group. (In practice, this
does not happen, because both the saving of prefabs and the copying to the
clipboard works with cloned elements, and cloning does not cover any group
associations. Alas, e.g. with hand-edited files, it would still be possible
to see groups in prefabs or the clipboard.)

As the prefabs or clipboard loading code is not prepared to deal with any
groups newly introduced by them (e.g. duplicates are not detected, the
observers such as the Groups toolbar are not properly updated, the command
history is not used, etc.), this change makes sure that prefabs and the
clipboard are always loaded without groups.