Map Editor: Keep the "auto-group entities" flag in the `MapDocumentT`, so that it can be toggled both via toolbars and menu items.
This change allows user code to learn the state of the "auto-group entities"
flag by a simple call to the new method `MapDocumentT::GetAutoGroupEntities()`
rather than awkwardly from the `OptionsBar_SelectionToolT` instance as before.

It also makes it possible to have a menu item (besides the checkbox in the
Selection tool's toolbar) to toggle this setting, and optionally to add a
proper `wxAuiToolBar` toggle button later.