Map Editor: Make sure that the Entity Hierarchy properly reflects the "true" selection.
After a selection click in the Entity Hierarchy, the "true" selection in the
map document may differ from what is indicated in the Entity Hierarchy control,
see the comments in the code for examples.

This change makes sure that the Entity Hierarchy properly reflects whatever has
become the "true" selection in the map document.