GameSys: Add new component `ComponentTargetT`.
This component significantly helps with writing better, reusable script code,
allowing entities that act in concert to be kept in a more flexible and
independent manner.

For example, with Target components the scripts that implement "open door" GUIs
no longer have to know the doors that they should act upon from a hard-coded
setting. Instead, they can now be flexibly reused with arbitrary door entities
as set in the Map Editor.