GameSys: Add frequently used attributes to the Mover component and update the DeathMatch sample behavior script `MoverBinary.lua` accordingly.
Most movers share a common set of attributes, such as the duration of a move.
This change adds a set of such attributes that are frequently and commonly used
to the Mover component, such that they can both be conveniently set in the
Map Editor and be readily accessed in the Mover's behavior script.

The `MoverBinary.lua` sample script has been updated accordingly, addressing
all of the newly added attributes.