ClipSys: Fix bogus assert failure in `CollisionModelStaticT` ctor.
If brushes have bevel planes, they have sides with no vertices at all. If the
last side in `m_BrushSides` was a side without vertices, in the last iteration
of loop

    unsigned long FirstVINr = 0;
    for (unsigned long SideNr = 0; SideNr < m_BrushSides.Size(); SideNr++)
    {
        m_BrushSides[SideNr].Vertices = &m_BrushSideVIs[FirstVINr];
        FirstVINr += m_BrushSides[SideNr].NrOfVertices;
    }

we already have `FirstVINr == m_BrushSideVIs.Size()`, triggering an
"out-of-bounds" assertion in the `m_BrushSideVIs[FirstVINr]` access.
This change fixes the issue.